Make types accessible for other modules
This makes it possible to easily mock the client. At the moment that is
a bit cumbersome as the types are internal.
