+++ livedns/snapshots.go
@@ -1,9 +1,8 @@
 package livedns
 
 import (
+	"github.com/go-gandi/go-gandi/types"
 	"time"
-
-	"github.com/go-gandi/go-gandi/internal/client"
 )
 
 // Snapshot represents a point in time record of a domain
@@ -23,7 +22,7 @@ func (g *LiveDNS) ListSnapshots(fqdn string) (snapshots []Snapshot, err error) {
 }
 
 // CreateSnapshot creates a snapshot for a domain
-func (g *LiveDNS) CreateSnapshot(fqdn string) (response client.StandardResponse, err error) {
+func (g *LiveDNS) CreateSnapshot(fqdn string) (response types.StandardResponse, err error) {
 	_, err = g.client.Post("domains/"+fqdn+"/snapshots", nil, &response)
 	return
 }
