+++ types/types.go
@@ -1,0 +1,19 @@
+package types
+
+// StandardResponse is a standard response
+type StandardResponse struct {
+	Code    int             `json:"code,omitempty"`
+	Message string          `json:"message,omitempty"`
+	UUID    string          `json:"uuid,omitempty"`
+	Object  string          `json:"object,omitempty"`
+	Cause   string          `json:"cause,omitempty"`
+	Status  string          `json:"status,omitempty"`
+	Errors  []StandardError `json:"errors,omitempty"`
+}
+
+// StandardError is embedded in a standard error
+type StandardError struct {
+	Location    string `json:"location"`
+	Name        string `json:"name"`
+	Description string `json:"description"`
+}
